+++ src/main/resources/META-INF/plugin.xml
@@ -33,10 +33,6 @@
                     anchor="bottom"
                     factoryClass="org.digma.intellij.plugin.toolwindow.recentactivity.DigmaBottomToolWindowFactory"/>
 
-        <projectService serviceImplementation="org.digma.intellij.plugin.toolwindow.recentactivity.RecentActivityCustomViewerWindowService"/>
-
-        <projectService serviceImplementation="org.digma.intellij.plugin.toolwindow.sidepane.InstallationWizardCustomViewerWindowService"/>
-
         <projectService serviceImplementation="org.digma.intellij.plugin.ui.ToolWindowShower"/>
 
         <projectService serviceImplementation="org.digma.intellij.plugin.service.EditorInteractionService"
